Refectoring: Additional person refactoring + initial population refactoring
